Respect ESLint extensions in deduplicating rules and config writing
The original ESLint configuration is now included in the `writeConversionResults` written file with a single `...`.

Additionally skipped outputting rules that match an existing ESLint plugin. If an extended ESLint plugin defines a rule with the same values as the resultant config, the config will skip printing it.
This necessitated expanding "failures" to allow a new `ConfigurationError` type of error for unfound ESLint configurations.

Renames "packages" in the configuration reading/converting logic areas to "plugins", as that's what ESLint calls them.