Allowed custom .js or .json file outputs with --config (#75)
As before, the output path defaults to `--eslintrc.js`. `--eslint` will default to `--config` which has that default.
The format of the output path depends on its file extension, with JavaScript for `.js` files and JSON otherwise.

The README.md list of flags is getting a little long so I added a quick manual table of contents. Will try to remember to split that into a separate PR...